+++ social_spam/vkontakte.py
@@ -16,6 +16,9 @@ class Vkontakte:
         self.vk = None
         self.message = None
         self.image = None
+        self.document = None
+        self.video = None
+        self.audio = None
 
     def connect_user(self, token: str) -> None:
         """
@@ -115,10 +118,42 @@ class Vkontakte:
             access_key = photo[0]['access_key']
             self.image = f'photo{owner_id}_{photo_id}_{access_key}'
 
+    def set_document(self, path: str) -> None:
+        """
+        Attach document to message
+
+        Args:
+            path (str): document path
+        """
+        if Path(path).is_file():
+            upload = vk_api.VkUpload(self.vk)
+            doc = upload.document_message(path)
+            owner_id = doc['doc']['owner_id']
+            doc_id = doc['doc']['id']
+            access_key = doc['doc']['access_key']
+            self.document = f'doc{owner_id}_{doc_id}_{access_key}'
+
+    def set_audio_message(self, path: str) -> None:
+        """
+        Attach audio message (voice message) to message
+
+        Args:
+            path (str): audio file path (mp3, ogg, etc.)
+        """
+        if Path(path).is_file():
+            upload = vk_api.VkUpload(self.vk)
+            audio = upload.audio_message(path)
+            owner_id = audio['audio_message']['owner_id']
+            audio_id = audio['audio_message']['id']
+            access_key = audio['audio_message']['access_key']
+            self.audio = f'audio_message{owner_id}_{audio_id}_{access_key}'
+
     def start_selective_spam(self,
                              chats: list,
                              message: str = None,
                              image: str = None,
+                             document: str = None,
+                             audio: str = None,
                              delay_time: float = 1
                              ) -> None:
         """
@@ -128,31 +163,46 @@ class Vkontakte:
             chats (list): ID of all users
             message (str): set a text message
             image (str): set image to message
+            document (str): set document to message
+            audio (str): set audio message to message
             delay_time (float): delay time between sending emails
         """
         if message is not None:
             self.set_message(message)
         if image is not None:
             self.set_image(image)
+        if document is not None:
+            self.set_document(document)
+        if audio is not None:
+            self.set_audio_message(audio)
+
+        # Build attachment string
+        attachment = None
+        if self.image:
+            attachment = self.image
+        elif self.document:
+            attachment = self.document
+        elif self.audio:
+            attachment = self.audio
 
         with alive_bar(len(chats), force_tty=True) as bar:
             for id in chats:
                 if str(abs(id)).isdigit():
-                    if self.image is None:
-                        self.vk.messages.send(user_id=abs(id), message=self.message, random_id=0)
-                        time.sleep(delay_time)
-                        bar()
-                    else:
+                    if attachment:
                         self.vk.messages.send(user_id=abs(id), message=self.message, random_id=0,
-                                              attachment=self.image)
-                        time.sleep(delay_time)
-                        bar()
+                                              attachment=attachment)
+                    else:
+                        self.vk.messages.send(user_id=abs(id), message=self.message, random_id=0)
+                    time.sleep(delay_time)
+                    bar()
                 else:
                     print(f'{id} is not ID')
 
     def start_all_spam(self,
                        message: str = None,
                        image: str = None,
+                       document: str = None,
+                       audio: str = None,
                        delay_time: float = 1
                        ) -> None:
         """
@@ -161,31 +211,46 @@ class Vkontakte:
         Args:
             message (str): set a text message
             image (str): set image to message
+            document (str): set document to message
+            audio (str): set audio message to message
             delay_time (float): delay time between sending emails
         """
         if message is not None:
             self.set_message(message)
         if image is not None:
             self.set_image(image)
+        if document is not None:
+            self.set_document(document)
+        if audio is not None:
+            self.set_audio_message(audio)
+
+        # Build attachment string
+        attachment = None
+        if self.image:
+            attachment = self.image
+        elif self.document:
+            attachment = self.document
+        elif self.audio:
+            attachment = self.audio
 
         chats = self.get_chats_id()
         with alive_bar(len(chats), force_tty=True) as bar:
             for chat in chats:
-                if self.image is None:
-                    self.vk.messages.send(user_id=abs(chat), message=self.message, random_id=0)
-                    time.sleep(delay_time)
-                    bar()
-                else:
+                if attachment:
                     self.vk.messages.send(user_id=abs(chat), message=self.message, random_id=0,
-                                          attachment=self.image)
-                    time.sleep(delay_time)
-                    bar()
+                                          attachment=attachment)
+                else:
+                    self.vk.messages.send(user_id=abs(chat), message=self.message, random_id=0)
+                time.sleep(delay_time)
+                bar()
 
     def start_bombing(self,
                       user_id: int,
                       amount: int,
                       message: str = None,
                       image: str = None,
+                      document: str = None,
+                      audio: str = None,
                       delay_time: float = 1
                       ) -> None:
         """
@@ -196,29 +261,44 @@ class Vkontakte:
              amount (int): number of sent messages
              message (str): set a text message
              image (str): set image to message
+             document (str): set document to message
+             audio (str): set audio message to message
              delay_time (float): delay time between sending emails
         """
         if message is not None:
             self.set_message(message)
         if image is not None:
             self.set_image(image)
+        if document is not None:
+            self.set_document(document)
+        if audio is not None:
+            self.set_audio_message(audio)
+
+        # Build attachment string
+        attachment = None
+        if self.image:
+            attachment = self.image
+        elif self.document:
+            attachment = self.document
+        elif self.audio:
+            attachment = self.audio
 
         with alive_bar(amount, force_tty=True) as bar:
             for i in range(amount):
-                if self.image is not None:
+                if attachment:
                     self.vk.messages.send(user_id=abs(user_id), message=self.message, random_id=0,
-                                          attachment=self.image)
-                    time.sleep(delay_time)
-                    bar()
+                                          attachment=attachment)
                 else:
                     self.vk.messages.send(user_id=abs(user_id), message=self.message, random_id=0)
-                    time.sleep(delay_time)
-                    bar()
+                time.sleep(delay_time)
+                bar()
 
     def send_message(self,
                      user_id: int,
                      message: str = None,
-                     image: str = None
+                     image: str = None,
+                     document: str = None,
+                     audio: str = None
                      ) -> None:
         """
         Normally send a message to a user
@@ -227,14 +307,59 @@ class Vkontakte:
             user_id (int): user ID
             message (str): set a text message
             image (str): set image to message
+            document (str): set document to message
+            audio (str): set audio message to message
         """
         if message is not None:
             self.set_message(message)
         if image is not None:
             self.set_image(image)
+        if document is not None:
+            self.set_document(document)
+        if audio is not None:
+            self.set_audio_message(audio)
 
-        if self.image is not None:
+        # Build attachment string
+        attachment = None
+        if self.image:
+            attachment = self.image
+        elif self.document:
+            attachment = self.document
+        elif self.audio:
+            attachment = self.audio
+
+        if attachment:
             self.vk.messages.send(user_id=abs(user_id), message=self.message, random_id=0,
-                                  attachment=self.image)
+                                  attachment=attachment)
         else:
             self.vk.messages.send(user_id=abs(user_id), message=self.message, random_id=0)
+
+    def get_friends(self) -> list:
+        """
+        Get list of friends
+
+        Return:
+            list: list of friends with their info (id, name, etc.)
+        """
+        try:
+            friends = self.vk.friends.get(fields='first_name,last_name,nickname')
+            return friends['items']
+        except Exception as e:
+            print(f"Error getting friends: {e}")
+            return []
+
+    def get_user_info(self, user_id: int) -> dict:
+        """
+        Get detailed information about a user
+
+        Args:
+            user_id (int): user ID
+        Return:
+            dict: user information
+        """
+        try:
+            user = self.vk.users.get(user_ids=user_id, fields='photo_max,city,country,bdate,contacts')[0]
+            return user
+        except Exception as e:
+            print(f"Error getting user info: {e}")
+            return None